projects/job-change-detection/requirements.txt
# Google Sheets
gspread>=5.12.0
google-auth-oauthlib>=1.2.0

# Async HTTP
aiohttp>=3.9.0

# Retry logic
tenacity>=8.2.0

# Environment variables
python-dotenv>=1.0.0

# SSL certificates
certifi>=2023.7.22
